requirements.txt
anthropic==0.28.0
fastapi==0.111.0
uvicorn==0.30.0
apscheduler==3.10.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
twilio==8.13.0
elevenlabs==1.2.0
requests==2.32.3
pytrends==4.9.2
feedparser==6.0.11
loguru==0.7.2
